+++ Core/source/StickyShotModifier.cpp
@@ -43,7 +43,7 @@ namespace bc
         {
             std::vector<Entity*> affectedEntities;
             se::Vector2 midPoint(0.0f, 0.0f);
-            for (int i = 0; i < (*this->entities)[CollisionGroup::Enemies].size(); ++i)
+            for (unsigned int i = 0; i < (*this->entities)[CollisionGroup::Enemies].size(); ++i)
             {
                 if (se::Math::Distance((*this->entities)[CollisionGroup::Enemies][i].getSprite().getPosition(), otherEntity->getSprite().getPosition()) <= this->radius)
                 {
@@ -52,9 +52,9 @@ namespace bc
                 }
             }
 
-            midPoint /= affectedEntities.size();
+            midPoint /= (float)affectedEntities.size();
 
-            for (int i = 0; i < affectedEntities.size(); ++i)
+            for (unsigned int i = 0; i < affectedEntities.size(); ++i)
             {
                 std::vector<Entity*> otherEntities = affectedEntities;
                 otherEntities.erase(std::find(otherEntities.begin(), otherEntities.end(), affectedEntities[i]));
@@ -64,7 +64,7 @@ namespace bc
                 affectedEntities[i]->modifiers.push_back(modifier);
             }
 
-            for (int i = 0; i < 50; ++i)
+            for (unsigned int i = 0; i < 50; ++i)
             {
                 std::vector<IModifier*> modifiers;
                 modifiers.push_back(new ParticleModifier(se::Vector2(rand() % 2001 - 1000, rand() % 2001 - 1000), se::Vector2(2.0f, 2.0f), rand() % 501 / 1000.0f));
